requirements.txt
openai-agents>=0.0.11
asyncio>=3.4.3
python-dotenv>=1.0.0
pillow>=11.2.1
numpy>=2.2.4
playwright==1.42.0
requests>=2.31.0
PyJWT>=2.11.0
